refactor(fundadores): split page into sub-components for LOC gate
Split FundadoresClient.tsx (440 LOC, +73% growth) into 5 sub-components
to satisfy audit-godmodule-loc.yml >20% growth gate:
- FundadoresFounderLetter
- FundadoresComparisonTable
- FundadoresDoNotBuy
- FundadoresGuarantee
- FundadoresLegalFooter

Also compress FundadoresFeatures.tsx (78 -> 42 LOC) by inlining
FEATURES literals onto single lines. No copy/visual changes.

Result:
- FundadoresClient.tsx: 440 -> 191 LOC (base 253; -24% growth)
- FundadoresFeatures.tsx: 78 -> 42 LOC (base 61; -31% growth)
- 5 new files all <100 LOC each

Tests: 11/11 pass in __tests__/fundadores/FundadoresClient.test.tsx
